Split state.js into focused modules (3169 -> 1502 lines)
- Extract roof auto-generation into roof-generation.js (AppState keeps thin delegates)
- Move geometry helpers into geometry-utils.js; drop state.js re-exports
- Extract section-catalog.js, serialization.js, model-ops.js, display-settings.js
- Deduplicate constructor/loadJSON defaults via createDefaultSettings/resetRuntimeState
- Table-driven patch sanitizers for updateSurface/updateMember
- Add state.select(kind, id) and revision counter
- Serialize element ids in toJSON (loadJSON stays compatible with id-less files);
  snapshot() now uses structuredClone
- Replace magic numbers with constants.js references
